Isolate WASM settings fragments in shadow DOM
Introduced tb-html-fragment custom element that attaches a shadow
root and injects fragment HTML inside it, replacing direct innerHTML
assignment into a light-DOM container.

Prevents WASM component settings pages from leaking styles/markup
into or colliding with the admin app's own DOM and CSS, enforcing
isolation at the injection point rather than relying on convention.
